==> R/01_load.qmd <==
---
title: "01_load"
format:
  html:
    embed-resources: true
editor: visual
execute: 
  echo: true
  message: false
  warning: false
---

## Loading libraries

```{r}
library("tidyverse")
library("here")
```

## Loading data

Checking if the data folder is there and if not, creates it and writes the loaded data file.

```{r}
project_root <- here()
data_dir <- here("data")
if (!dir.exists(data_dir)) dir.create(data_dir)

input_file <- here("_raw", "kidney_disease.csv")
output_file <- here("data", "01_data_load.csv")

read_csv(input_file) |> write_csv(output_file)
```
