==> R/04_describe.qmd <==
---
title: "04_describe"
format:
  html:
    embed-resources: true
editor: visual
execute: 
  echo: true
  message: false
  warning: false
---

## Load libraries and data

```{r load_libraries}
library("tidyverse")
library("table1")
library("gt")
```

```{r load_data}
kidney_merged_GFR <- read_rds("../data/03_data_augment.rds")
```

## General information

This data was gathered during a 2-month period in India, following individuals with or without Chronic Kidney Disease. This is the main classification; in addition, the dataset registers 25 other features (e.g. cell count, appetite, hemoglobin) while placing individuals in correct age groups, the information about sex is missing.

Overview of the variables can be found here:

```{r variable_table}
ckd_overview <- tribble(
  ~Variable, ~Description, ~Type, ~Units,
  "age", "Patient age", "Numeric", "years",
  "blood_pressure", "Measured blood pressure", "Numeric", "mm/Hg",
  "blood_glucose_random", "Blood glucose (random)", "Numeric", "mg/dL",
  "blood_urea", "Urea concentration in blood", "Numeric", "mg/dL",
  "serum_creatinine", "Creatinine in blood", "Numeric", "mg/dL",
  "sodium", "Blood Sodium level", "Numeric", "mEq/L",
  "potassium", "Blood Potassium level", "Numeric", "mEq/L",
  "hemoglobin", "Hemoglobin level", "Numeric", "g/dL",
  "packed_cell_volume", "Percentage of RBC volume", "Numeric", "%",
  "white_blood_cell_count", "White blood cell count", "Numeric", "cells/cumm",
  "red_blood_cell_count", "Red blood cell count", "Numeric", "millions/cmm",
  "GFR_x", "Glomerular Filtration Rate", "Numeric", "ml/minute",

  "age_group", "Age group", "Categorical", "9 age groups",
  "specific_gravity", "Urine concentration level", "Categorical", "{1.005–1.025}",
  "albumin", "Albumin level in urine", "Categorical (ordinal)", "0–5 scale",
  "sugar", "Sugar level in urine", "Categorical (ordinal)", "0–5 scale",
  "red_blood_cells", "Red blood cells in urine", "Categorical", "normal / abnormal",
  "pus_cells", "Pus cells presence in urine", "Categorical", "normal / abnormal",
  "pus_cell_clumps", "Pus cell clumps", "Categorical", "present / notpresent",
  "bacteria", "Bacterial presence in urine", "Categorical", "present / notpresent",
  "hypertension", "Hypertension status", "Categorical", "yes / no",
  "diabetes_mellitus", "Diabetes status", "Categorical", "yes / no",
  "coronary_artery_disease", "Coronary artery disease status", "Categorical", "yes / no",
  "appetite", "Reported appetite", "Categorical", "good / poor",
  "pedal_edema", "Swelling in legs/feet", "Categorical", "yes / no",
  "anemia", "Anemia status", "Categorical", "yes / no",
  "classification", "CKD diagnosis", "Categorical", "ckd / notckd",
  "ckd_stage", "Stage of CKD", "Categorical", "1–5 scale"
)

ckd_overview |>
  arrange(Type) |> 
  gt(groupname_col = "Type") |>
  tab_options(
    table.width = pct(100),
    row_group.as_column = TRUE
  ) |>
  tab_header(
    title = "Chronic Kidney Disease Dataset: Variable Overview"
  )


```

After the addition of classification in the previous step, the final dataset has 400 observations and a total of 31 variables.

```{r get_nRC}
dim(kidney_merged_GFR)
```

## Inspecting NA values

The percentage of NAs in each column is summarized here:

```{r calculate_NAs}

nas <- kidney_merged_GFR |> 
     summarise(across(everything(), ~mean(is.na(.)) * 100))

nas |> 
  select(-red_blood_cells) |> 
  unlist() |> 
  mean() 
```

The category with the most missing values is the classification of red blood cells, with 38% missing values. Apart from this category, the average of NAs is around 8%. We decided not to delete all rows with missing values, as this would result in a drop of observations from 400 to 158. Therefore, we will handle them individually in the analysis.

```{r get_rows}
nrow(kidney_merged_GFR)
nrow(kidney_merged_GFR |> drop_na())
```

## Table 1

This table provides us with descriptive statistics of the main characteristics of the data stratified by classification (ckd/notckd). We see that there are 250 individuals diagnosed with chronic kidney disease, and 150 not diagnosed, with mean age 54.5 and 46.5 for the two groups respectively.

```{r generate_table1}
table1(
  ~ age +
    blood_pressure +
    specific_gravity +
    albumin +
    sugar +
    serum_creatinine +
    blood_urea +
    GFR_average +
    red_blood_cells +
    pus_cells +
    pus_cell_clumps +
    bacteria +
    hemoglobin +
    packed_cell_volume +
    diabetes_mellitus +
    hypertension +
    coronary_artery_disease +
    appetite +
    anemia +
    pedal_edema
    |
    classification,
  data = kidney_merged_GFR
)

```

## Age distribution

Individuals in the dataset include people from 2 to 90 years old. Table1 shows the median, however, we can also visualise the age distribution for a better overview. The most prevalent groups are people aged from 41 to 70 years old, the majority of which has been diagnosed with chronic kidney disease. The youngest and the oldest age groups are represented only by diagnosed individuals.

```{r count_people_in_age_group}
kidney_filter <- kidney_merged_GFR |>  
  drop_na(age_group) 

ggplot(kidney_filter, aes(age_group, fill = classification)) +
  geom_bar(
    position = position_dodge(preserve = "single")) +
  theme_minimal() +
  labs(title="Kidney disease distribution",
       x = ("Age Group"),
       y = ("Count"),
       fill = "Classification"
       )

ggsave(
    filename = "../results/04_age_dist.png"
  )
```

## CDK stage distribution

We categorized CKD into stages based on the blood filtration efficiency. The plot above shows how many individuals are in which stage - Normal, Mild, Mild to Moderate, Moderate to Severe, Severe or Failure - the last one often requiring dialysis or transplant.

As expected, we see that most of the individuals not diagnosed with CDK fall within the Normal to Mild category. On the contrary, among individuals with CDK, the majority of individuals belong to advanced stages of the disease, with GFR from 30 to 40, and over 60 people suffer from kidney failure.

```{r heatmap}
#| fig-width: 8
#| fig-height: 6
stage_labs <- c("Normal", 
                "Mild", 
                "Mild to moderate", 
                "Moderate to severe",
                "Severe", 
                "Failure")

kidney_heatmap <- kidney_merged_GFR |> 
  subset(!is.na(cdk_stage)) |> 
  group_by(cdk_stage, classification) |> 
  summarise(count = n(), .groups = "drop")

ggplot(kidney_heatmap, aes(x = cdk_stage, y = classification, fill = count)) +
  geom_tile(color = "white") +  
  geom_text(aes(label = count), color = "darkgrey") +
  scale_fill_gradient(low = "white", high = "orange") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 25, hjust = 1)) +
  labs(title = "CKD Stage Distribution",
       x = NULL,
       y = "Classification",
       fill = "Count") +
  scale_x_discrete(labels = stage_labs)

ggsave(
    filename = "../results/04_CKD_stage_dist.png"
  )
```
