==> R/05_analysis_1.qmd <==
---
title: "05_analysis_1"
format:
  html:
    embed-resources: true
editor: visual
execute: 
  echo: true
  message: false
  warning: false
---

# Variables analysis

## Loading libraries

```{r}
library("PerformanceAnalytics")
library("randomForest")
library("tidyverse")
library("caret")
library("dplyr")
```

## Setting seed for reproducibility

```{r}
set.seed(123)
```

## Loading augmented data

```{r}
kidney_augmented <- read_rds("../data/03_data_augment.rds")
```

## Correlation plot

The corr plot shows the correlation between all the numeric variables in the dataset, from there we can inspect if there is correlation between variables and how strong.

```{r}
kidney_augmented |>
  select(age, blood_pressure, blood_glucose_random, 
         blood_urea, serum_creatinine, sodium, potassium, 
         hemoglobin, packed_cell_volume, white_blood_cell_count, 
         red_blood_cell_count, GFR_male, GFR_female, GFR_average) |>
  chart.Correlation(histogram = TRUE, pch = 19)

```

## Random forest analysis

Using random forest we are inspecting the variable importance, to further select our variables. This also includes factor variables.

```{r}
data <- kidney_augmented |> 
  na.omit() |>
  select(-c(GFR_female, GFR_male, GFR_average, cdk_stage, specific_gravity))

idx <- data |>
  pull(classification) |>
  createDataPartition(p = 0.6, list = F)

train <- data[idx,]
test <- data[-idx,]

rf1 <- train |> 
  randomForest(classification ~ ., data = _, importance = T, ntree = 600)

rf1$err.rate[,1] |>
  plot(type='l')

```

### Visualizing the confusion matrix

This is used to see the model performance on the test set.

```{r}
truth <- test |>
  pull(classification)
pred  <- rf1 |>
  predict(test)

confusionMatrix(truth, pred)

```

### Variables importance plot

This plot is used to see how much each variable is influencing the accuracy of the model.

```{r}
varImpPlot(rf1)

png("../results/05_rf1_varImp.png", width = 800, height = 1200, type = "cairo")
varImpPlot(rf1)
dev.off()

```

### Repeating random forest analysis with only important variables

```{r}
kidney_augmented <- read_rds("../data/03_data_augment.rds")

data <- kidney_augmented |> 
  na.omit() |>
  select(-c(GFR_female, GFR_male, GFR_average, cdk_stage, specific_gravity))

idx <- data |> 
  pull(classification) |>
  createDataPartition(p = 0.6, list = F)

train <- data[idx,]
test <- data[-idx,]

rf2 <- train |> 
  randomForest(classification ~ packed_cell_volume + hemoglobin +
                 serum_creatinine + red_blood_cell_count + albumin + hypertension + diabetes_mellitus,
               data = _, 
               importance = T, 
               ntree = 600)

rf2$err.rate[,1] |>
  plot(type='l')

```

### Visualizing the confusion matrix

This is used to see the model performance on the test set.

```{r}
truth <- test |>
  pull(classification)
pred  <- rf2 |>
  predict(test)

confusionMatrix(truth, pred)
```

### Variables importance plot

This plot is used to see how much each variable is influencing the accuracy of the model.

```{r}
varImpPlot(rf2)

png("../results/05_rf2_varImp.png", width = 800, height = 1200, type = "cairo")
varImpPlot(rf2)
dev.off()
```
