==> R/06_analysis_2.qmd <==
---
title: "06_analysis_2"
format:
  html:
    embed-resources: true
editor: visual
execute: 
  echo: true
  message: false
  warning: false
---

## Loading libraries

```{r}
library("tidyverse")
library("DescTools")
```

## Loading augmented data

```{r}
kidney_merged_GFR <- read_rds("../data/03_data_augment.rds")
```

## Prepare variables for Analysis

Here we make sure that the key categorical variables are treated as factors and that the CKD stages have an ordered structure.

```{r}
kidney_analysis <- kidney_merged_GFR |>
mutate(
  classification = as.factor(classification),
  hypertension = as.factor(hypertension),
  diabetes_mellitus = as.factor(diabetes_mellitus),
  coronary_artery_disease = as.factor(coronary_artery_disease),
  anemia = as.factor(anemia),
  pedal_edema = as.factor(pedal_edema),
  cdk_stage = factor(
    cdk_stage,
    levels = c("stage_1", "stage_2", "stage_3a", "stage_3b", "stage_4", "stage_5"),
    ordered = TRUE
    )
  )

kidney_analysis |>
  write_rds("../data/04_data_analysis.rds")
```

## Question 1: Is kidney function (GFR) lower in CKD patients?

We expect individuals with a CKD diagnosis to have lower estimated glomerular filtration rate (GFR).

```{r}
kidney_analysis |>
  drop_na(GFR_average, classification) |>
  group_by(classification) |>
  summarise(
    n = n(),
    mean_GFR = mean(GFR_average),
    median_GFR = median(GFR_average),
    sd_GFR = sd(GFR_average)
    )

```

Visualize GFR by CKD classification

```{r}
#| fig-width: 8
#| fig-height: 5

kidney_analysis |>
  drop_na(GFR_average, classification) |>
  ggplot(aes(x = classification, y = GFR_average, fill = classification)) +
  geom_boxplot() +
  theme_minimal() +
  theme(legend.position="none") +
  labs(
    title = "Estimated GFR by CKD classification",
    x = "Chronic kidney disease outcome",
    y = "Estimated GFR (ml/min)") +
  scale_x_discrete(
    labels = c("ckd" = "Disease", "notckd" = "No Disease")) +
  scale_fill_manual(
    values = c("ckd" = "#C44536", "notckd" = "#6A994E"))

ggsave(
    filename = "../results/06_gfr_by_ckd.png"
  )

```

## Question 2: How are CKD stages distributed between diagnosed and non-diagnosed individuals?

Here we look at how the GFR-based CKD stages are distributed in people with and without a CKD diagnosis.

```{r}
kidney_analysis |>
  drop_na(cdk_stage, classification) |>
  count(cdk_stage, classification) |>
  group_by(classification) |>
  mutate(prop = n / sum(n))

```

```{r}
#| fig-width: 8
#| fig-height: 5

stage_labs <- c("Normal", 
                "Mild", 
                "Mild to moderate", 
                "Moderate to severe",
                "Severe", 
                "Failure")

kidney_analysis |>
  drop_na(cdk_stage, classification) |>
  ggplot(aes(x = cdk_stage, fill = classification)) +
  geom_bar(position = "fill") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(
    title = "Distribution of GFR-based CKD stages by classification",
    x = NULL,
    y = "Proportion within classification group"
    ) +
  scale_y_continuous(labels = scales::percent_format()) +
  scale_fill_manual(
    values = c("ckd" = "#C44536", "notckd" = "#6A994E"),
    labels = c("ckd" = "CKD diagnosis", "notckd" = "No CKD diagnosis")) +
  scale_x_discrete(labels = stage_labs)

ggsave(
    filename = "../results/06_CKD_dist.png"
  )

```

## Question 3: Are hypertension and diabetes more common among CKD patients?

Hypertension and diabetes are known risk factors for CKD. Here we compare their prevalence between groups.

```{r}
kidney_analysis |>
  drop_na(c(hypertension, classification)) |>
  count(classification, hypertension) |>
  group_by(classification) |>
  mutate(prop = n / sum(n))

```

Visualization

```{r}
#| fig-width: 8
#| fig-height: 5

kidney_analysis |>
  drop_na(c(hypertension, classification)) |>
  ggplot(aes(x = classification, fill = hypertension)) +
  geom_bar(
    position = position_dodge(preserve = "single")) +   
  theme_minimal() +
  labs(
    title = "Hypertension by CKD classification",
    fill = "Hypertension",
    x = "Chronic kidney disease outcome",
    y = "Number of people") +
  scale_x_discrete(
    labels = c("ckd" = "Disease", "notckd" = "No Disease")) +
  scale_fill_manual(
    values = c("yes" = "#C44536", "no" = "#6A994E"),
    labels = c("yes" = "Diagnosed", "no" = "Not diagnosed"))

ggsave(
    filename = "../results/06_CKD_hypertension.png"
  )
```

## Diabetes vs Classification

```{r}
kidney_analysis |>
  drop_na(diabetes_mellitus, classification) |>
  count(classification, diabetes_mellitus) |>
  group_by(classification) |>
  mutate(prop = n / sum(n))

```

```{r}
#| fig-width: 8
#| fig-height: 5

kidney_analysis |>
  drop_na(diabetes_mellitus, classification) |>
  ggplot(aes(x = classification, fill = diabetes_mellitus)) +
  geom_bar(position = position_dodge(preserve = "single")) +
  theme_minimal() +
  labs(
    title = "Diabetes mellitus by CKD classification",
    fill = "Diabetes mellitus",
    x = "Chronic kidney disease outcome",
    y = "Number of people") +
  scale_x_discrete(
    labels = c("ckd" = "Disease", "notckd" = "No Disease")) +
  scale_fill_manual(
    values = c("yes" = "#C44536", "no" = "#6A994E"),
    labels = c("yes" = "Diagnosed", "no" = "Not diagnosed"))

ggsave(
    filename = "../results/06_CKD_diabetes.png"
  )

```

## Investigating hypertension and diabetes mellitus

Here we investigate why the two variables seem equally discriminatory but one is labelled as non important by the random forest analysis.

```{r}
tab <- kidney_analysis |>
  xtabs(~ hypertension + diabetes_mellitus, data = _)

tab
tab |>
  chisq.test()

V <- CramerV(tab) |>
  round(3)

sprintf("Cramér's V = %.3f", V)

```

It looks like the variables are strongly correlated and that could explain what is happening.

## Final Plot

Setting the data to long format

```{r}
kidney_long <- kidney_analysis |>
  select(
    classification,
    albumin,
    serum_creatinine,
    hemoglobin,
    packed_cell_volume,
    red_blood_cell_count) |>
  pivot_longer(
    cols = -classification,
    names_to = "variable",
    values_to = "value") |>
  drop_na(value)

```

Adjusting for label positions

```{r}
label_pos <- kidney_long |>
  group_by(variable) |>
  summarise(
    y_pos = max(value, na.rm = TRUE) * 1.15,
    .groups = "drop")
```

Count observations per group and join label positions

```{r}
n_labels <- kidney_long |>
  count(variable, classification, name = "n") |>
  left_join(label_pos, by = "variable")

```

Final Plot visualization

The levels of key predictors - albumin, hemoglobin, cell volume, cell count of red blood, and creatinine are chosen for the final visualisation. RBC count and related packed cell volume are showing lower levels in individuals who are diagnosed with CKD. This is due to the kidneys not producing enough of hormone erythropoietin, responsible for signalling the creation of red blood cells in the bone marrow. This shortage can lead to anemia. This strongly correlates with hemoglobin levels and the expected result is shown in the plot - lower hemoglobin levels for diagnosed individuals. Creatinine levels, also used for GFR calculation, follow the opposite pattern. Creatinine, as a waste product from muscle and protein breakdown, is filtered by the kidneys. When their function is compromised, this product is found in the blood. As seen in the plot, healthy individuals show levels close to zero while elevated levels are recorded in diagnosed individuals. Lastly, albumin found in urine is an early indicator of kidney damage, as this essential protein should only stay in the blood. This is a clear indicator of a kidney filtration issue that can usually predict the progression of the disease. Non-diagnosed individuals show all scores 0, while diagnosed individuals range from 0 to 3.

```{r}
#| fig-width: 10
#| fig-height: 4

ggplot(
  kidney_long,
  aes(x = classification, y = value, fill = classification)) +
  geom_boxplot() +
  geom_text(
    data = n_labels,
    aes(y = y_pos, label = paste0("n = ", n)), size = 3) +
  facet_wrap(
    ~ variable,
    scales = "free_y",
    nrow = 1,
    labeller = as_labeller(
      c(
        albumin = "Albumin (score)",
        hemoglobin = "Hemoglobin (g/dL)",
        packed_cell_volume = "PCV (%)",
        red_blood_cell_count = "RBC count (millions/mcL)",
        serum_creatinine = "Creatinine (mg/dL)"
        )
      )
    ) +
  theme_minimal() +
  labs(
    title = "Key continuous predictors by CKD classification",
    fill = "Classification:",
    x = "",
    y = "") +
  theme(
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    legend.position = "bottom") +
  scale_fill_manual(
    values = c("ckd" = "#C44536", "notckd" = "#6A994E"),
    labels = c("ckd" = "Diagnosed", "notckd" = "Not diagnosed"))

ggsave(
    filename = "../results/06_key_predictors.png"
  )

```
